pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.3",
    "django-extensions>=4.1",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-spectacular[sidecar]>=0.28.0",
    "pillow>=11.2.1",
    "pydotplus>=2.0.2",
]
